Basic User Interface created for main, danger mode, pass code
created activity DangerMode that has SAFE button, UI is functional but
simple. Further enhancements are expected. No other interaction beyond
the UI was created. Click and LongClick for the SAFE button were
enabled, still researching on how to enable a "button released"
function, I believe it's involved in the LongClick function. This is
for https://github.com/peterbee/cs390final/issues/4

created initial layouts for classes: main, DangerMode, pass code

pass code has no activity attached to it yet. I am still researching
how to utilize it within the app. The UI is very rough and early in
design. This is for https://github.com/peterbee/cs390final/issues/3

Initial plans for the main activity is to enable the logic to determine
if a user is new or not and then to launch the appropriate activites
for the user. No ticket for this issue at the moment.